requirements.txt
aiogram==3.13.1
ollama==0.3.3
python-dotenv==1.0.1
